Extracted game logic into its own file
Everything used to be hacked into wstty.js -- which was supposed to be
only the terminal control class. But since this project is becoming less
of a hack and more of a proper screensaver, we decided to properly modularize things.

wstty.js now contains only terminal related things.
game_control.js is responsible to reason about nethack.